Create csv-to-keychain-access
For anyone reading this, this imports the passwords to your Keychain Access. From there, you can visit the website on Safari and copy and paste the password once to remember it for future ease of use.
Format of file should be csv with each line being URL,username,password without quotes. This Applescript runs through and add each code to Keychain Access. Allows for duplicate but may have some issues and stop halfway. To resolve this, run your csv file a few times till you are sure all the passwords are added. Thank you.